Airflow/requirements.txt
apache-airflow-providers-google
pandas
faker
pandas_gbq
dbt-core
dbt-bigquery
airflow-dbt-python
google-genai
google-cloud-bigquery